Implemented knight travails function
Implemented the knight travails a bit out of scope of what was asked.

createBoard factory function: returns an object that has an array
representing a gameboard and a method that removes a space from it, with
the intention that method gets called whenever a space is visited on the
gameboard.

knightsTravails: Is a function that iteratively moves a knight from the
startPoint to the endPoint. Within the function, the knight travels as
long as there are spaces available to visit. Once all spaces have been
visited, we return the shortest array in the array with all paths that
lead to the endPoint.

This is out of scope, because the assignment intended for a graph to be
the data structure used. Not a 2d array.

 Changes to be committed:
	modified:   knightsTravails.js