Added spaceNode & board factory functions
A new file has been added that appropiately starts to tackle the problem
provided by TOP here: https://www.theodinproject.com/lessons/foundations-fundamentals-part-5#intermediateadvanced-array-magic

Given this project is about executing against the possible moves a
knight can make, the spaceNode FF is based on only the moves a knight
can make.

spaceNode: returns a node that has...
-- Coordinates
-- Pointers to other spaceNodes based on the 8 potential moves a knight
can make.

gameBoard: This FF creates a graph in which every space on a chess board
is represented in a spaceNode and the node points to every possible
space a knight can jump too.

 Changes to be committed:
	new file:   knightTravails2.js